@@ -7,7 +7,7 @@
     "test": "jest --coverage --env=node",
     "snyk-test": "snyk test",
     "link": "npm run build && cd ./dist && npm link",
-    "lint": "eslint ./src --ext .ts --output-file ./eslint-report.json --format json",
+    "lint": "eslint ./src --output-file ./eslint-report.json --format json",
     "postbuild": "bash package.sh",
     "schematic:list": "npm run build && schematics ./dist/schematics/collection.json:list employee --path=\"./dist/\" --openApiJsonUrl=https://im-wa-empo-nrcrn.azurewebsites.net/swagger/v1/swagger.json --dry-run=false --force --verbose --appPrefix=nrcrn",
     "schematic:module:bld": "npm run build && schematics ./dist/schematics/collection.json:module building --path=\"./dist/\" --openApiJsonUrl=https://im-wa-unto-nrcrn.azurewebsites.net/swagger/v1/swagger.json --dry-run=false --force --verbose --appPrefix=nrcrn",
@@ -39,19 +39,22 @@
     "@angular-devkit/schematics-cli": "^18.1.4",
     "@angular/compiler": "^18.1.4",
     "@angular/core": "^18.1.4",
+    "@eslint/js": "^9.8.0",
     "@types/jest": "^29.5.12",
     "@types/lodash": "^4.17.7",
     "@types/node": "^18.19.43",
     "@types/pluralize": "^0.0.33",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.8.0",
-    "eslint": "^8.57.0",
+    "@typescript-eslint/eslint-plugin": "^8.0.1",
+    "@typescript-eslint/parser": "^8.0.1",
+    "eslint": "^9.8.0",
     "eslint-config-prettier": "^9.1.0",
+    "globals": "^15.9.0",
     "jest": "^29.7.0",
     "jest-junit": "^16.0.0",
     "schematics": "^0.2.0",
     "snyk": "^1.1292.2",
     "ts-jest": "^29.2.4",
-    "typescript": "^5.5.4"
+    "typescript": "^5.5.4",
+    "typescript-eslint": "^8.0.1"
   }
 }
